refactor blueprints (6/6): extract command_deck blueprint
Final blueprint extraction. ~25 routes under /command-deck/* plus
the two _huyang_build_* helpers move to blueprints/command_deck.py
(~775 lines).

Routes moved:
- Dashboard, projects CRUD (/command-deck/, /projects/, /projects/new,
  /projects/<slug>/, /update, /delete)
- Blocks (add, update, update-title, delete, reorder)
- Checklist (toggle, add, delete)
- Project tasks (add, complete, delete)
- File uploads (upload, delete)
- Promote-task
- Huyang chat (chat, history, clear)
- PIN gate (verify-pin)

Module constants: PRIVATE_PROJECTS_PIN, ANTHROPIC_API_KEY redefined
locally from os.environ.

Intra-blueprint url_for references (cd_projects, cd_project) stay as
unqualified names — Flask resolves them inside the same blueprint
without the prefix.

Smoke-tested: /command-deck/, /command-deck/projects/, /command-deck/
verify-pin (403 for wrong PIN, by design); /publish, /today/, /below-deck,
/ani/ping, /cockpit/mode all 200 (full cross-blueprint pass).

Refactor of app.py complete: from 2606 lines (pre-refactor) to ~141
lines now (imports, env-var constants, six blueprint registrations,
runner). All ~65 routes live in their respective blueprint files.
